+++ relay-server/src/processing/utils/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "processing")]
+pub mod store;
