+++ tests/integration/test_trace_metrics.py
@@ -10,16 +10,6 @@ from .asserts import time_within_delta
 TEST_CONFIG = {
     "outcomes": {
         "emit_outcomes": True,
-        "batch_size": 1,
-        "batch_interval": 1,
-        "aggregator": {
-            "bucket_interval": 1,
-            "flush_interval": 1,
-        },
-    },
-    "aggregator": {
-        "bucket_interval": 1,
-        "initial_delay": 0,
     },
 }
